Some notes on imaging processing using PIL.
Perhaps it can be used to discard all black (or near black) images.

Ideally you'd want to have a list of black images and a list of non-black images and just create a video based on the non-black images. It may also be sensible to keep the black images (and make a video of them) incase the thresholds are not sensible.